Change misleading reference to ctb.thread_index().x into warp.thread_rank()
In warp shuffle algorithms, the x axis of the thread index is equal to
the thread's rank in the warp, as the x axis is set to 32 to make
grouping of threads into warps easier. This may be misleading when
reading the code, so changed it into warp.thread_rank() instead.

Add default distribution type to for n_to_m optimized shuffle. Change it
to triangle from none.