Compile multiple kernels based on max shifts, right matrices and left rows
Before, we only compiled kernels for different values of left rows. All
kernels then used the same amount of registers, even if they really
required just a few for small values of max shifts and max right
matrices.

Now the maximum number of shifts and right matrices is also given to the
kernel as tempalte argument, which means there are different kernels for
different values of these arguments. This allows us to measure the
different requirements on registers and resulting effects of occupation.

Unfortunately compilation times are now around 42 minutes, which will
further suffer when we add work distribution. There is nothing we can do
with that, apart from testing with smaller limits and only doing the big
compilation when measuring.