Implement n_to_m shuffle with multirow optimization
Implement n_to_m shuffle with multirow optimization. This is basically
the ultimate n_to_m shuffle algorithm. We could add work distribution,
but it would have to be reimplemented so that it could work with workers
processing tasks from different shifts, which is not currently possible.

The implementation does not use any local memory, so it works as
intended.

This implementation includes the debug prints for future reference when
debugging. These were used to catch a bug in the main loop where when
reading the thread_left_bottom, multiplying with right_val and adding to
sum, we were indexing using l_mat * NUM_LEFT_MATS + l_row instead of
l_mat * NUM_LEFT_ROWS + l_row. This resulted in out of bound memory
access, which in this case lead to NaN values instead of crashes, which
is what threw me off.

Otherwise the algorithm is basically just an amalgamation of the n_to_m
with work_dist with work_dist stripped out, and the n_to_mn with
multirow.

Renamed NUM_RIGHT_ROWS in n_to_mn, which did not fit with the other
places this value is passed from or to, where it is named
NUM_SHIFTS_PER_RIGHT_MATRIX, or NUM_SHIFTS_PER_MAT for short, as there
is only a single left matrix and multiple right matrices.

Moved the zeroing out of the d_result_ array using cudaMemset before
every kernel start, as it needs to be zeroed out before every run of the
algorithm. This is to enable us to run multiple iterations of the kernel
while still computing correct results.