Move the algorithm argument limits to cmake as cache variables
Before, the limits were set by hardcoded constants in the source code.
As we need to change these between release and debug builds and we may
need to change these when building on some higher end cards with more
registers, we now pass these limits from the cmake configuration as
preprocessor macros.

The limits can be set using cmake gui or using the -D falg during cmake
config run.

Renamed cudaStream arguments to cuda_stream, to match the snake case we
use everywhere else.

Changed the manually written switches in old algorithms to the recursive
template dispatch so that these work with different limit values, which
can now be set during cmake configuration.

Added host side dispatch to all algorithms. Before, each algorithm
geenrated just a single kernel based on the limit values. This did not
allow us to test the effects of differing occupancy between the
configuration limits.

Now each algorithm generates one kernel for each combination of values
up to the limit, and each run the implementation is chosen based on the
algorithm arguments. This significantly increases the compilation times
and executable size, but allows us to test what we need.