Rename files to unify the naming scheme of algorithms
We need to unify and simplify the naming scheme of algorithms. First the
multirow and multileft distinction, which is about rows, clashes with
multiright, which means multiple right matrices. We now introduce
multirow_right and multirow_both, together with multimat_right and
multimat_both.

We leave the n_to_m name in the implementation specific for n_to_m, as
it is only usable for n_to_m.

n_to_mn, which is actually one_to_many, is renamed accordingly.