Rename block_y_size and rows_per_block to warps_per_thread_block
Rename and standardize the block_y_size and rows_per_block algorithm
arguments into warps_per_thread_block, which is what the argument really
means.

Rename shifts_per_block to shifts_per_thread_block, which also basically
determines the number of warps per thread block, but as this is in the
warp_per_shift algorithm, I think it is better to give it the
shifts_per_thread_block name.