Rename orig versions to local_mem, rename warp_shuffle to shuffle_multimat_right
Rename the orig versions of multirow_both and multimat_both to local_mem
versions, as they are the original unfixed versions which use local
memory. The local_mem is more descriptive for the future when we forget
what was the problem in the original version.

Rename the basic warp_shuffle implementation to a more descriptive and
uniform shuffle_multimat_right, as that is the proper name using our
naming scheme.

Rename shift_per_block to block_per_shift.

Rename the compile time constants to match.

Change validation benchmark to match the changes to algorithm names,
arguments and to add new algorithms.