Use few of the implementations for n_to_mn alg, fix bugs
Utilize few of our shuffle implementations for n_to_mn computation type.
This is achieved by starting multiple kernels in different streams, one
for each left matrix, so that they all compute in parallel.

The major problem was that the original algorithm and all our
subsequently created infrastructure works with the right matrices
interleaved, so that we have m blocks of n right matrices, where in each
block there is one matrix for each left matrix.

To use our one to many implementations, we need to reorder this to have
n blocks of m matrices, where all right matrices to be computed with
given left matrix are contiguous in the right data array.

This functionality was added to data_array, which can now load and store
matrices from and to the files interleaved. It can also create new
data_array containing interleaved matrices in memory.

To deinterleave, we provide the group size n, i.e. the number of left
matrices. To interleave again, we provide group size m, i.e. the number
of groups in the interleaved ordering.

This also has to be done when loading validation results, as those need
to also be deinterleaved to compare with the results of our
implementations. Same when computing the valid results on CPU, where we
furthermore need to interleave again the right matrices.

We also need to interleave the results again when storing them to .csv,
so that they match other results of n_to_mn.

This change necessitated some additions to the cross_corr base class. We
added new algorithm step, store_results, which is implemented by
store_results_impl virtual method. The default implementation provided
by the base class preserves prior behavior, which prevents us from
needing to change all other implementations. Only n_to_mn which use the
one_to_many implementations need to override this.

We also needed to route valid result loading through the leaf classes.
This is done by the new load_valid_results virtual method. Again,
default implementation preserves prior behavior.

While we were at it, we did some simplification. The get_valid_result is
now renamed to compute_valid_results which better reflects what it does.

Move the load_matrix_array... methods to matrix.hpp, which is where they
make more sense than in cross_corr.hpp.

Add cuda streams to the one_to_many implementations used for n_to_mn.

Add fft_alg mixin class, which allows us to identify fft algs during
compile time easily and to move result normalization to the algorithm
classes themselfs intead of having it in main.

Add the new algorithms to validation.

Fix unknown algorithm output, which used to contain just a series of
commas instead of comma delimited list of available algorithms. This was
because we were just initializing a vector with the size of the set
instead of copying the set into the vector, which ment the vector
contained the correct amount, but of empty strings.

Fix invalid assert in n_to_m_multirow_both_multimat_both implementation.
Zero shifts per warp is valid if it is completely outside the output
matrix due to the fixed size of thread blocks.

Fix the limitation of cpu cross correlation implementations, which
needed to have the exact same allocators. When interleaving the right
inputs again, we do not want to allocate more pinned memory, so we have
a different allocator for left and right inputs. This allows us to still
pass the inputs to cpu implementations.

Fix exception when no arguments are provided to the cross executable.
Now we print help message instead.

This commit contains commented out debugging main implementation for the
data_array interleaving and deinterleaving. This is left for future
reference and will be removed in the next commit.