Add previous failure skipping when benchmarking
Before, when using the continue switch to continue after previous run,
we detected the previous runs by their results. But there are no results
for failed runs, so these were ran again and failed again, which is
unnecessarily slow. Now we parse the failure_log.yml file and if
continuing, we check that the given run did not fail previously,
otherwise we skip it.

Each group has its own failure log. We split the failures by input
index, then run index and finally full run name. The problem was that
input_index and run index do not identify each run uniquely, as runs
generated from a definition with a range of argument values all share
their index. This is a bad design, but we do not have the time to change
it, so we have to identify the runs using their full names.

Simplified Logger, so that it does not control the lifetime of the
failure_log file. It just requires the standard log output and failure
log output TextIO streams in the constructor and uses them for output.