Add last kernel launch argument tracking
Add last kernel launch argument tracking, so that we can include this
information when an algorithm run fails. All failures should be due to
invalid arguments, i.e. too many threads, too many blocks or too much
shared memory. All of these values will now be included in the error
message when algorithm run fails, which allows us to "debug" the failure
much more easily.

We have new kernel_args class, which is derived in each algorithm
implementation to also include other arguments, such as the template
args. All derived classes are singletons, and after each launch, the
arguments are written into the singleton and pointer to that singleton
is stored as the last kernel launch arguments.

Add type as static constexpr property to each row_distribution class.
This is used when recording kernel launch arguments.

Fix CMakeLists.txt not respecting the changes to the algorithm option.
We were just using the default value instead of passing in the value set
by the user in CMake Cache.

Fix validation benchmark definition, which used invalid argument
left_rows_per_iteration_ which does nothing.

Add missing assert in naive_shuffle_multimat_right for right matrices
per thread 0. It is checked in the run function, so it should never be 0
in the dispatch function.