Remove shared memory from multirow implementations
We used shared memory as we needed to pass offset to the output array to
the comppute_row_groups device function. But it turns out that the
offset is compile time constant, so now we pass it as a template
argument instead.

This allows us to pass in a reference to local array, which is placed
into registers. We then utilize the offset from template argument to
index into the array.

The only tricky thing is ordering in the sum array when computing
multiple matrices. We need to exclude the first N overlaps from ALL
matrices, so we have to order the shifts as shift_0_lmat_0_rmat_0,
shift_0_lmat_0_rmat_1 ... shift_0_lmat_1_rmat_0 ...
shift_1_lmat_0_rmat_0 for the multimat_both, shift_0_rmat_0,
shift_0_rmat_1, shift_1_rmat_0 ... for multimat_right.

Otherwise the change is pretty straight forward, more of a refactoring.