Update README, add sample app, add example benchmark, switch to SciPy as default validator
Update README to the final version, to be included in the submitted file
and final version of the repository. Describe everything such as
building and running the CUDA C++ app, running the benchmarks, running
visualization, contents of the repository, how to take one
implementation and use it in sample app etc.

Swtich default benchmarking type to None

Add sample_app, which showcases how to take an cross-correlation
implementation from our code and use it in an application. All source
files other than main.cpp are just symlinks to the global src directory
files.